Extract metrics into a separate module
